requirements.txt
fastapi
uvicorn
openai
pytesseract
python-multipart
Pillow
beautifulsoup4
requests
python-dotenv
